resume_parser/requirements.txt
flask>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=0.8.11
requests>=2.31.0
scikit-learn>=1.3.0
nltk>=3.8.1
